preloaded core data try
