--- weatherApp/App/Data/CityLocationServiceProtocol.swift
@@ -1,13 +1,0 @@
-//
-//  CityLocationServiceProtocol.swift
-//  weatherApp
-//
-//  Created by Dario Nikolic on 04/09/2020.
-//  Copyright © 2020 Dario Nikolic. All rights reserved.
-//
-
-import Foundation
-
-protocol CityLocationServiceProtocol {
-    
-}
