+++ weatherApp/Features/SearchLocation/UI/LocationSearchPresenter.swift
@@ -7,20 +7,31 @@
 //
 
 import RxSwift
+import RxCocoa
 
 class LocationSearchPresenter {
     
+    private let cityListdisposeBag: DisposeBag = DisposeBag()
     private let weatherRepository: WeatherRepository
     private let navigationService: NavigationService
+    private var cityList: BehaviorRelay<[City]>
     
     init(weatherRepository: WeatherRepository,navigationService: NavigationService) {
         self.weatherRepository = weatherRepository
         self.navigationService = navigationService
+        self.cityList = BehaviorRelay<[City]>(value: [])
+        
+        bindCityList()
     }
     
-    func fetch(query: String) {
-        let result = weatherRepository.fetchCities(query: query)
-        print(result)
+    func queryLocationSearch(with query: String) {
+        
+    }
+    
+    private func bindCityList() {
+        weatherRepository.fetchCities()
+        .bind(to: cityList)
+        .disposed(by: cityListdisposeBag)
     }
     
 }
